Cover to uint64

+++ helper/resource/id.go
@@ -32,11 +32,11 @@ func UniqueId() string {
 // in the order of their creation, even across multiple terraform executions, as
 // long as the clock is not turned back between calls.
 func PrefixedUniqueId(prefix string) string {
-	timestamp := strconv.FormatUint(time.Now().Unix(), 36)
+	timestamp := strconv.FormatUint(uint64(time.Now().Unix()), 36)
 
 	idMutex.Lock()
 	defer idMutex.Unlock()
 	idCounter++
-	id := strconv.FormatUint(idCounter, 36)
+	id := strconv.FormatUint(uint64(idCounter), 36)
 	return fmt.Sprintf("%s%s%s", prefix, timestamp, id)
 }
